Move PlayerStateSystem in correct path
